+++ Renderers/IRenderer.hpp
@@ -34,9 +34,9 @@ namespace renderer {
             const raytracer::Vector2<float> &position,
             const unsigned size) = 0;
         virtual void drawObject(const objects::IObject &object,
-            const raytracer::Vector3<int> &color) = 0;
+            const raytracer::Vector3<double> &color) = 0;
         virtual void drawPixel(const raytracer::Vector2<int> &position,
-            const raytracer::Vector3<int> &color) = 0;
+            const raytracer::Vector3<double> &color) = 0;
 
         virtual void render() = 0;
         virtual void clear() = 0;
